Rename demo project to todo-app
