+++ backend/src/main/java/com/example/todo/TodoApplication.java
@@ -1,15 +1,15 @@
-package com.example.demo;
+package com.example.todo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan(basePackages = {"com.example.demo.controller", "com.example.demo.service", "com.example.demo.config"})
-public class DemoApplication {
+@ComponentScan("com.example.todo")
+public class TodoApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(TodoApplication.class, args);
 	}
 
 }
