+++ backend/src/main/java/com/example/todo/config/WebConfig.java
@@ -1,13 +1,19 @@
-package com.example.demo.config;
+package com.example.todo.config;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import com.example.demo.constants.UrlConst;
+import com.example.todo.constants.UrlConst;
+import com.example.todo.interceptor.LoggingInterceptor;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
+
+    @Autowired
+    private LoggingInterceptor loggingInterceptor;
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
@@ -18,5 +24,10 @@ public class WebConfig implements WebMvcConfigurer {
                 .exposedHeaders(UrlConst.CORS_EXPOSED_HEADERS)
                 .allowCredentials(Boolean.parseBoolean(UrlConst.CORS_ALLOWED_CREDENTIALS))
                 .maxAge(Long.parseLong(UrlConst.CORS_MAX_AGE));
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(loggingInterceptor);
     }
 }
