+++ backend/src/main/java/com/example/todo/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto;
+package com.example.todo.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
