+++ backend/src/test/java/com/example/todo/TodoApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.demo;
+package com.example.todo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class TodoApplicationTests {
 
 	@Test
 	void contextLoads() {
